move wrap to imgui utilities
