+++ unit-tests/utilities/imgui/test-wrap.cpp
@@ -1,12 +1,12 @@
 // License: Apache 2.0. See LICENSE file in root directory.
 // Copyright(c) 2020 Intel Corporation. All Rights Reserved.
 
-//#cmake:add-file ../../../common/utilities/string/wrap.h
-//#cmake:add-file ../../../common/utilities/string/wrap.cpp
+//#cmake:add-file ../../../common/utilities/imgui/wrap.h
+//#cmake:add-file ../../../common/utilities/imgui/wrap.cpp
 //#cmake:add-file ../../../third-party/imgui/imgui.h
 
 #include "common.h"
-#include "../../../common/utilities/string/wrap.h"
+#include "../../../common/utilities/imgui/wrap.h"
 #include "../../../third-party/imgui/imgui.h"
 
 namespace ImGui {
@@ -32,7 +32,7 @@ ImVec2 CalcTextSize( const char * text,
 }
 }  // namespace ImGui
 
-using namespace utilities::string;
+using namespace utilities::imgui;
 
 TEST_CASE( "wrap-text", "[string]" )
 {
